Allow for custom per-fork CNAMEs and keep history of pages
When the action is run in fork, it will not override CNAME that
is already there. Also we are going to keep the whole history of
gh-pages rather than amending the last commit - there will be
more traceability and the commits will make more sense.

For each gh-pages deployment we add the GITHUB_SHA of the commit
from the main page that triggered the build.